pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rewired-index"
version = "0.1.0"
description = "Rewired Index: AI-powered 5-layer investment framework with signal monitoring"
requires-python = ">=3.11"
dependencies = [
    "yfinance>=1.2.0",
    "fredapi>=0.5.2",
    "google-genai>=1.0.0",
    "python-telegram-bot>=22.0",
    "pandas>=2.2",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "click>=8.1",
    "rich>=13.0",
    "schedule>=1.2",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]
gui = [
    "nicegui>=2.0",
]

[project.scripts]
rewired = "rewired.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
